nvme-cmds: split header according to spec grouping
The nvme-cmds header is currently oversized and difficult to maintain.
Split the header into logical sections following the groupings defined
in the NVMe specification.


merge with adebd7b0879b ("nvme-cmds: split header according to spec grouping")
