+++ libnvme/src/meson.build
@@ -37,6 +37,11 @@ headers = [
     'nvme/lib.h',
     'nvme/linux.h',
     'nvme/mem.h',
+    'nvme/nvme-cmds-base.h',
+    'nvme/nvme-cmds-fabrics.h',
+    'nvme/nvme-cmds-mi.h',
+    'nvme/nvme-cmds-nvm.h',
+    'nvme/nvme-cmds-zns.h',
     'nvme/nvme-cmds.h',
     'nvme/nvme-types.h',
     'nvme/tree.h',
