+++ libnvme/test/meson.build
@@ -253,6 +253,11 @@ foreach hdr : [
     'nbft',
     'nbft-types',
     'nvme-cmds',
+    'nvme-cmds-base',
+    'nvme-cmds-fabrics',
+    'nvme-cmds-mi',
+    'nvme-cmds-nvm',
+    'nvme-cmds-zns',
     'nvme-types',
     'tree',
     'util',
